q5.cpp

+++ level1/q5.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main()
 {
-    int age = 25;
+  short int age = 25;
 
     cout << "After 5 years you will be "
          << age + 5 << " years old\n";
